fix code smells for sonarcloud

+++ util/concurrent/include/ConcurrentHashMap.hpp
@@ -15,7 +15,7 @@ public:
         wrLock = rdwrLock->getWriteLock();
     };
 
-    int size() {
+    size_t size() {
         AutoLock l(rdLock);
         return mMap->size();
     }
